+++ src/main/kotlin/com/github/reschoene/model/TaskList.kt
@@ -1,5 +1,5 @@
 package com.github.reschoene.model
 
 data class TaskList (var id: String = "", var name: String = ""){
-    var taskCount: Long = 0L
+    var taskCount: Int = 0
 }
